+++ PERUSTARS/PERUSTARS/AtEventManagement/Domain/Services/Participant/IParticipantQueryService.cs
@@ -5,9 +5,9 @@ namespace PERUSTARS.AtEventManagement.Domain.Services.Participant
 {
     public interface IParticipantQueryService
     {
-        public PERUSTARS.AtEventManagement.Domain.Model.Aggregates.Participant getParticipantById(int id);
-        public IEnumerable<PERUSTARS.AtEventManagement.Domain.Model.Aggregates.Participant>  getParticipantByHobbyistId(int id);
-        public IEnumerable<PERUSTARS.AtEventManagement.Domain.Model.Aggregates.Participant> getParticipantByEventId(int id);
+        public PERUSTARS.AtEventManagement.Domain.Model.Aggregates.Participant getParticipantById(long id);
+        public IEnumerable<PERUSTARS.AtEventManagement.Domain.Model.Aggregates.Participant>  getParticipantByHobbyistId(long id);
+        public IEnumerable<PERUSTARS.AtEventManagement.Domain.Model.Aggregates.Participant> getParticipantByEventId(long id);
         public IEnumerable<PERUSTARS.AtEventManagement.Domain.Model.Aggregates.Participant> GetParticipants();
     }
 }
